refactor: rename LoginForm component to Login in App.jsx
