+++ src/App.jsx
@@ -3,7 +3,7 @@ import { AuthProvider } from './context/AuthContext';
 import { CartProvider } from './context/CartContext';
 
 // Auth Components
-import LoginForm from './components/auth/LoginForm';
+import LoginForm from './components/presentational/auth/LoginForm';  // Fix path
 import RegisterForm from './components/auth/RegisterForm';
 import ForgotPasswordForm from './components/auth/ForgotPasswordForm';
 import ResetPasswordForm from './components/auth/ResetPasswordForm';
